Add lib file for gui functions
